Fixed a compiler warning
size is of type off_t which is signed int64 on 64bit Mac.

Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>

+++ src/3rdparty/btree/src/btree.cpp
@@ -3927,7 +3927,7 @@ btree_dump(struct btree *bt)
                         bt->meta.entries,
                         bt->meta.leaf_pages,
                         bt->meta.branch_pages,
-                        bt->size,
+                        (long)bt->size,
                         bt->head.psize);
                 btree_dump_tree(bt, bt->meta.root, 0);
         } else {
